Moved all classes to separate files within the classes directory
