+++ src/classes/Ability.ts
@@ -1,0 +1,9 @@
+export class Ability {
+  constructor(name, description, energy, value, multiplier) {
+    this.name = name;
+    this.description = description;
+    this.energy = energy;
+    this.value = value;
+    this.multiplier = multiplier;
+  }
+}
