+++ src/classes/Class.ts
@@ -1,0 +1,5 @@
+export class Class {
+  constructor(name) {
+    this.name = name;
+  }
+}
