+++ src/classes/User.ts
@@ -1,0 +1,13 @@
+export class User {
+  username;
+  password;
+
+  constructor(username, password) {
+    this.username = username;
+    this.password = password;
+  }
+
+  verifyPassword(value) {
+    return value === this.password;
+  }
+}
