+++ src/classes/command/Command.ts
@@ -1,0 +1,10 @@
+export class Command {
+  constructor(commandKey, commandFunction, displayName, displayed) {
+    this.commandKey = commandKey;
+    this.commandFunction = commandFunction;
+    this.displayName = displayName;
+  }
+  executeCommand() {
+    this.commandFunction();
+  }
+}
