+++ src/views/statusView.js
